PKIRA-173 - Users can see a list of other registrants within a domain
Note: this only applies if the user's registration is approved.

Modification of web service to also return a list of approved registrants for the domain if the registration itself is approved.
Modification of portal code to map the webservice result and show it on the screen.
Stylesheet and formatting modification to display the results as nowrap.