PKIRA-174 - The application is translated in FR/EN/DE/NL
Addition of i18n to the application (portal) based on Seam translations.
Correct language is passed to eID IdP and eID DSS.
Menu is changed to let the user enter the language.

Note: list of languages is hard-coded in menu.xhtml to avoid the invalid label for FR in Seam.